% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_run_to_list.R
\name{add_run_to_list}
\alias{add_run_to_list}
\title{Add a new AVENIO run to patient overview}
\usage{
add_run_to_list(master_list, Directory)
}
\arguments{
\item{master_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}

\item{Directory}{\code{Character} string with the full path to the directory on
the Synology where the data is located.}
}
\value{
A \code{list} of \code{data.frames} with similar structure to \code{master_list}.
During the analysis the number of patients and NGS runs present in the
dataset before and after the NGS run is printed. If the list successfully
has been updated the \code{.rds} file on the Synology containing the list is
automatically updated.
}
\description{
This function takes the main list of NGS data and adds the information from
a new run to list. If the patient ID/CPR is not present in the main list
the patient is added to the list. If a patient ID/CPR is present in the list
the new NGS data is added under that patient and all the BAM files for that
patient is manually checked for mutations.
}
\examples{
 master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
 test_path <- "//Synology_m1/Synology_folder/AVENIO/AVENIO_results/Plasma-ANei_Hsi__lGmaJ1k_Ji_d3O"
 add_run_to_list(master_list = master,
                 Directory = test_path)
 @export
}
