% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simple_output.R
\name{create_simple_output}
\alias{create_simple_output}
\title{Get a simple overview for a patient}
\usage{
create_simple_output(
  df_list,
  CPR,
  synonymous = TRUE,
  synology_path = "//Synology_m1/Synology_folder/AVENIO/"
)
}
\arguments{
\item{df_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}

\item{CPR}{\code{Character} string of the unique CPR number or other unique
number for the patient of interest. Available CPR numbers can be viewed
using names(\code{df_list})}

\item{synonymous}{\code{Boolean} determining whether synonymous mutations are kept
in the output. If \code{TRUE} (default) synonymous mutations are kept in the
output. If \code{FALSE} the synonymous mutations are excluded.}

\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}
}
\value{
A \code{list} of length = 2. The first entry ("Patient CPR:") shows the
patient of interest CPR/ID number. The second entry ("Results") is a \code{tibble}
with 14 variables. Each row corresponds to a mutation identified in any
sample. Important variables include sample_index which shows the project the
patient is connected to, the name of the patient in that project and the
date of of the sample collection. Sample.ID is the name of the sample in the
AVENIO system. Sample_note is the note specifying what type of blood
sample the sample is in relation to treatment (BL, Tx, BC, or Unknown).
Material specifies what type of material has been analyzed (cfDNA, BC,
cfChIP, size-selection, etc.) The Flags variable shows if the mutation only
has been found in the BAM file and not in the AVENIO output, and/or if the
mutation also has been found in a BC sample if such sample has been run for
the patient. If a sample has been analyzed but no mutation was identified
\code{NA} is in all relevant variables.
}
\description{
This function takes the main list of NGS data for each patient
located at "//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds"
and a CPR number (or other unique patient number used for the patient) and
creates a simple overview of the mutations identified.
}
\examples{
 master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
 names(master)
 create_simple_output(df_list = master,
                      CPR = "1234567890",
                      synonymous = FALSE)
}
