% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_AVENIO_runs.R
\name{explore_AVENIO_runs}
\alias{explore_AVENIO_runs}
\title{Explores the .xlsx file with the run information}
\usage{
explore_AVENIO_runs(
  Info = NULL,
  silent = FALSE,
  synology_path = "//Synology_m1/Synology_folder/AVENIO/"
)
}
\arguments{
\item{Info}{\code{Character} string with the name of the information that should.
be extracted. Possible options can be viewed with explore_AVENIO_runs_Info().
If \code{NULL} (default) all information is returned as a \code{list}.}

\item{silent}{\code{Boolean} determining information messages should be displayed
If \code{FALSE} (default) messages are displayed. If \code{TRUE} the messages are
not printed.}

\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}
}
\value{
A \code{list} with different information on the AVENIO_runs.xslx
(\code{Info} = \code{NULL}) or the specific object as determined by \code{Info}.
}
\description{
This function explores the AVENIO_runs.xlsx file and performs various tests
on the information and links the information to the list of mutations
for each patient
}
\examples{
results <- explore_AVENIO_runs()
results$Total_entries
explore_AVENIO_runs(Info = "Required")
}
