% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_AVENIO_runs_Info.R
\name{explore_AVENIO_runs_Info}
\alias{explore_AVENIO_runs_Info}
\title{Highlights the parameters to be investigated from explore_AVENIO_runs()}
\usage{
explore_AVENIO_runs_Info(
  synology_path = "//Synology_m1/Synology_folder/AVENIO/"
)
}
\arguments{
\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}
}
\value{
A \code{data.frame} with two variables containing the name and the
description of the entries in the explore_AVENIO_runs() output
}
\description{
This functions creates an overview of the different contents in the list
exported by explore_AVENIO_runs()
}
\examples{
explore_AVENIO_runs_Info()
}
