% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_project.R
\name{extract_project}
\alias{extract_project}
\title{Get a simple overview for an entire project}
\usage{
extract_project(df_list, project, synonymous = TRUE)
}
\arguments{
\item{df_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}

\item{project}{\code{Character} string of the project name as it appears in
AVENIO_runs.xlsx. Accepted project names are available with
\code{result_stats(Info = "Projectstats")}}

\item{synonymous}{\code{Boolean} determining whether synonymous mutations are kept
in the output. If \code{TRUE} (default) synonymous mutations are kept in the
output. If \code{FALSE} the synonymous mutations are excluded.}
}
\value{
A \code{tibble}
with 15 variables. Each row corresponds to a mutation identified in any
sample. Important variables include sample_index which shows the project the
patient is connected to, the name of the patient in that project and the
date of of the sample collection. CPR which is the CPR number of the patient,
Sample.ID is the name of the sample in the
AVENIO system. Sample_note is the note specifying what type of blood
sample the sample is in relation to treatment (BL, Tx, BC, or Unknown).
Material specifies what type of material has been analyzed (cfDNA, BC,
cfChIP, size-selection, etc.) The Flags variable shows if the mutation only
has been found in the BAM file and not in the AVENIO output, and/or if the
mutation also has been found in a BC sample if such sample has been run for
the patient. If a sample has been analyzed but no mutation was identified
\code{NA} is in all relevant variables.
}
\description{
This function takes the main list of NGS data for each patient
located at "//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds"
and a project ID and creates a simple overview of the mutations identified
for the patients in that project
}
\examples{
 master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
 result_stats(Info = "Projectstats")
 extract_project(df_list = master,
                 project = "Pembrolizumab",
                 synonymous = FALSE)
}
