% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_analyses.R
\name{included_analyses}
\alias{included_analyses}
\title{Get an overview of the analyses included in the dataset}
\usage{
included_analyses(master_list)
}
\arguments{
\item{master_list}{The \code{list} of \code{data.frames} which contains all information
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}
}
\value{
A \code{list} of length = 2. The first entry ("Overview:")
is a \code{tibble} with three variables giving an overview of the number of
patients (n) from each NGS run included in the dataset. The second entry is a
\code{list} of \code{data.frames} where each \code{data.frame} represents the samples from
the specific NGS run. Each \code{data.frame} is named according to the Analysis.ID
}
\description{
This function takes the main dataset and creates an overview of how many
samples from each NGS run is included in the dataset
}
\examples{
master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
analyses <- included_analyses(master)
analyses[[1]]
names(analyses[[2]])
analyses[[2]][["AbpTmAnr_h9DPL1skMvyje5U"]]
}
