% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_run.R
\name{remove_run}
\alias{remove_run}
\title{Remove all results for a specific sequencing run}
\usage{
remove_run(
  run,
  master_list,
  synology_path = "//Synology_m1/Synology_folder/AVENIO/",
  save = FALSE,
  save_as = paste0(synology_path, "AVENIO_results_patients.rds")
)
}
\arguments{
\item{run}{\code{Character} or \code{Character} \code{vector} of run_IDs you want
removed from the results}

\item{master_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}

\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}

\item{save}{\code{Boolean} Indicating whether the output without the specified
\code{run} should be automatically saved as
"//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds".
Default = \code{FALSE}}

\item{save_as}{Full path to the file you want to save the output as.
Default = paste0(synology_path,"AVENIO_results_patients.rds")}
}
\value{
A \code{list} of \code{data.frames} with similar structure to \code{master_list}.
but with the run_ID defined with \code{run} removed from the results
}
\description{
Remove results for a run_ID in cases of e.g., wrong entries in
'Avenio_runs.xlsx', errors during sequencing or the results are misleading
}
\examples{
master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
#Removing a single run from the results automatically
rm_single_run <- "EXAMPLE_RUN123"
remove_run(run = rm_single_run,
           master_list = master,
           save = TRUE)

#Removing several runs but without saving the results
rm_several_runs <- c("EXAMPLE_RUN123",
                     "EXAMPLE_RUN456",
                     "EXAMPLE_RUN789")
remove_run(run = rm_several_runs,
           master_list = master,
           save = FALSE)
}
