% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sample_index.R
\name{remove_sample_index}
\alias{remove_sample_index}
\title{Remove all results for a sample_index}
\usage{
remove_sample_index(
  samples,
  master_list,
  synology_path = "//Synology_m1/Synology_folder/AVENIO/",
  save = FALSE,
  save_as = paste0(synology_path, "AVENIO_results_patients.rds")
)
}
\arguments{
\item{samples}{\code{Character} or \code{Character} \code{vector} of sample_index you want
removed from the results}

\item{master_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}

\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}

\item{save}{\code{Boolean} Indicating whether the output without the specified
\code{samples} should be automatically saved as
"//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds".
Default = \code{FALSE}}

\item{save_as}{Full path to the file you want to save the output as.
Default = paste0(synology_path,"AVENIO_results_patients.rds")}
}
\value{
A \code{list} of \code{data.frames} with similar structure to \code{master_list}.
but with the sample_index defined with \code{samples} removed from the results
}
\description{
Remove results for a sample_index in cases of e.g., wrong entries in
'Avenio_runs.xlsx', errors during sequencing or the results are misleading
}
\examples{
master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
#Removing a single sample from the results automatically
rm_single_sample <- "PROJECT_PATIENT_260217"
remove_sample_index(samples = rm_single_sample,
                    master_list = master,
                    save = TRUE)

#Removing several samples but without saving the results
rm_several_samples <- c("PROJECT_PATIENT1_260217",
                        "PROJECT_PATIENT2_260217",
                        "PROJECT_PATIENT2_251023",
                        "PROJECT_PATIENT3_250701")
remove_sample_index(samples = rm_several_samples,
                    master_list = master,
                    save = FALSE)
}
