% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_stats.R
\name{result_stats}
\alias{result_stats}
\title{Generates an overview different stats in the dataset}
\usage{
result_stats(Info = NULL, silent = FALSE)
}
\arguments{
\item{Info}{\code{Character} string with the name of the information that should.
be extracted. Possible options can be viewed with result_stats_Info().
If \code{NULL} (default) all information is returned as a \code{list}.}

\item{silent}{\code{Boolean} determining information messages should be displayed
If \code{FALSE} (default) messages are displayed. If \code{TRUE} the messages are
not printed.}
}
\value{
A \code{list} with different stats on AVENIO_results_patients.rds
(\code{Info} = \code{NULL}) or the specific object as determined by \code{Info}.
}
\description{
This function takes the AVENIO_results_patients.rds file and generates
different explorative stats on the dataset
}
\examples{
results <- result_stats()
results$All_mutations
explore_AVENIO_runs(Info = "Basestats")
}
