% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_stats_Info.R
\name{result_stats_Info}
\alias{result_stats_Info}
\title{Highlights the parameters to be investigated from result_stats()}
\usage{
result_stats_Info(synology_path = "//Synology_m1/Synology_folder/AVENIO/")
}
\arguments{
\item{synology_path}{\code{Character} string with the full path to the directory
containing AVENIO_runs.xlsx & AVENIO_keys.rds. Default is
"//Synology_m1/Synology_folder/AVENIO/"}
}
\value{
A \code{data.frame} with two variables containing the name and the
description of the entries in the result_stats() output
}
\description{
This functions creates an overview of the different contents in the list
exported by result_stats()
}
\examples{
result_stats_Info()
}
