% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_frames.R
\name{unlist_frames}
\alias{unlist_frames}
\title{Create one big data.frame from list of patient data.frames}
\usage{
unlist_frames(master_list)
}
\arguments{
\item{master_list}{The \code{list} of \code{data.frames} which contains all information on
on the patients. This is read using
readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")}
}
\value{
A \code{data.frame} with the all samples
}
\description{
This function takes the main list of data.frames containing the mutation
information and creates a single data.frame with the same information,
where each individual is concatenated
}
\examples{
master <- readRDS("//Synology_m1/Synology_folder/AVENIO/AVENIO_results_patients.rds")
master_df <- unlist_frames(master_list = master)
print(master_df)
}
